Split asr.py into smaller modules.
