+++ ci/test_flake8.sh
@@ -8,8 +8,6 @@ set -euo pipefail
 flake8_black_list="\
 espnet/asr/asr_utils.py
 espnet/asr/chainer_backend/asr.py
-espnet/asr/pytorch_backend/asr.py
-espnet/bin/asr_enhance.py
 espnet/lm/chainer_backend/extlm.py
 espnet/lm/chainer_backend/lm.py
 espnet/lm/lm_utils.py
@@ -57,7 +55,6 @@ espnet/utils/fill_missing_args.py
 espnet/utils/io_utils.py
 espnet/utils/spec_augment.py
 espnet/utils/training/batchfy.py
-espnet/utils/training/evaluator.py
 espnet/utils/training/iterators.py
 espnet/utils/training/tensorboard_logger.py
 espnet/utils/training/train_utils.py
