+++ espnet/asr/pytorch_backend/train.py
@@ -1,27 +1,20 @@
-#!/usr/bin/env python3
+# Copyright 2020 The ESPnet Authors.
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
-"""
-This script is used for multi-speaker speech recognition.
+"""Backend module for `asr_train.py`."""
 
-Copyright 2017 Johns Hopkins University (Shinji Watanabe)
- Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
-"""
 import json
 import logging
 import os
 
-# chainer related
 from chainer import training
 from chainer.training import extensions
-from itertools import zip_longest as zip_longest
-import numpy as np
 from tensorboardX import SummaryWriter
 import torch
 
-from espnet.asr.asr_mix_utils import add_results_to_json
 from espnet.asr.asr_utils import adadelta_eps_decay
-
 from espnet.asr.asr_utils import CompareValueTrigger
+from espnet.asr.asr_utils import format_mulenc_args
 from espnet.asr.asr_utils import get_model_conf
 from espnet.asr.asr_utils import restore_snapshot
 from espnet.asr.asr_utils import snapshot_object
@@ -30,11 +23,12 @@ from espnet.asr.asr_utils import torch_resume
 from espnet.asr.asr_utils import torch_snapshot
 from espnet.asr.pytorch_backend.asr import CustomEvaluator
 from espnet.asr.pytorch_backend.asr import CustomUpdater
-from espnet.asr.pytorch_backend.asr_init import load_trained_model
-import espnet.lm.pytorch_backend.extlm as extlm_pytorch
+from espnet.asr.pytorch_backend.asr_init import freeze_modules
+from espnet.asr.pytorch_backend.asr_init import load_trained_modules
+from espnet.asr.pytorch_backend.converter import CustomConverter
+from espnet.asr.pytorch_backend.converter import CustomConverterMulEnc
 from espnet.nets.asr_interface import ASRInterface
-from espnet.nets.pytorch_backend.e2e_asr_mix import pad_list
-import espnet.nets.pytorch_backend.lm.default as lm_pytorch
+from espnet.nets.pytorch_backend.lm import default
 from espnet.utils.dataset import ChainerDataLoader
 from espnet.utils.dataset import TransformDataset
 from espnet.utils.deterministic_utils import set_deterministic_pytorch
@@ -46,86 +40,6 @@ from espnet.utils.training.tensorboard_logger import TensorboardLogger
 from espnet.utils.training.train_utils import check_early_stop
 from espnet.utils.training.train_utils import set_early_stop
 
-import matplotlib
-
-matplotlib.use("Agg")
-
-
-class CustomConverter(object):
-    """Custom batch converter for Pytorch.
-
-    Args:
-        subsampling_factor (int): The subsampling factor.
-        dtype (torch.dtype): Data type to convert.
-
-    """
-
-    def __init__(self, subsampling_factor=1, dtype=torch.float32, num_spkrs=2):
-        """Initialize the converter."""
-        self.subsampling_factor = subsampling_factor
-        self.ignore_id = -1
-        self.dtype = dtype
-        self.num_spkrs = num_spkrs
-
-    def __call__(self, batch, device=torch.device("cpu")):
-        """Transform a batch and send it to a device.
-
-        Args:
-            batch (list(tuple(str, dict[str, dict[str, Any]]))): The batch to transform.
-            device (torch.device): The device to send to.
-
-        Returns:
-            tuple(torch.Tensor, torch.Tensor, torch.Tensor): Transformed batch.
-
-        """
-        # batch should be located in list
-        assert len(batch) == 1
-        xs, ys = batch[0][0], batch[0][-self.num_spkrs :]
-
-        # perform subsampling
-        if self.subsampling_factor > 1:
-            xs = [x[:: self.subsampling_factor, :] for x in xs]
-
-        # get batch of lengths of input sequences
-        ilens = np.array([x.shape[0] for x in xs])
-
-        # perform padding and convert to tensor
-        # currently only support real number
-        if xs[0].dtype.kind == "c":
-            xs_pad_real = pad_list(
-                [torch.from_numpy(x.real).float() for x in xs], 0
-            ).to(device, dtype=self.dtype)
-            xs_pad_imag = pad_list(
-                [torch.from_numpy(x.imag).float() for x in xs], 0
-            ).to(device, dtype=self.dtype)
-            # Note(kamo):
-            # {'real': ..., 'imag': ...} will be changed to ComplexTensor in E2E.
-            # Don't create ComplexTensor and give it to E2E here
-            # because torch.nn.DataParallel can't handle it.
-            xs_pad = {"real": xs_pad_real, "imag": xs_pad_imag}
-        else:
-            xs_pad = pad_list([torch.from_numpy(x).float() for x in xs], 0).to(
-                device, dtype=self.dtype
-            )
-
-        ilens = torch.from_numpy(ilens).to(device)
-        if not isinstance(ys[0], np.ndarray):
-            ys_pad = []
-            for i in range(len(ys)):  # speakers
-                ys_pad += [torch.from_numpy(y).long() for y in ys[i]]
-            ys_pad = pad_list(ys_pad, self.ignore_id)
-            ys_pad = (
-                ys_pad.view(self.num_spkrs, -1, ys_pad.size(1))
-                .transpose(0, 1)
-                .to(device)
-            )  # (B, num_spkrs, Tmax)
-        else:
-            ys_pad = pad_list(
-                [torch.from_numpy(y).long() for y in ys], self.ignore_id
-            ).to(device)
-
-        return xs_pad, ilens, ys_pad
-
 
 def train(args):
     """Train with the given args.
@@ -135,6 +49,8 @@ def train(args):
 
     """
     set_deterministic_pytorch(args)
+    if args.num_encs > 1:
+        args = format_mulenc_args(args)
 
     # check cuda availability
     if not torch.cuda.is_available():
@@ -144,13 +60,27 @@ def train(args):
     with open(args.valid_json, "rb") as f:
         valid_json = json.load(f)["utts"]
     utts = list(valid_json.keys())
-    idim = int(valid_json[utts[0]]["input"][0]["shape"][-1])
+    idim_list = [
+        int(valid_json[utts[0]]["input"][i]["shape"][-1]) for i in range(args.num_encs)
+    ]
     odim = int(valid_json[utts[0]]["output"][0]["shape"][-1])
-    logging.info("#input dims : " + str(idim))
+    if hasattr(args, "decoder_mode") and args.decoder_mode == "maskctc":
+        odim += 1  # for the <mask> token
+    for i in range(args.num_encs):
+        logging.info("stream{}: input dims : {}".format(i + 1, idim_list[i]))
     logging.info("#output dims: " + str(odim))
 
     # specify attention, CTC, hybrid mode
-    if args.mtlalpha == 1.0:
+    if "transducer" in args.model_module:
+        if (
+            getattr(args, "etype", False) == "transformer"
+            or getattr(args, "dtype", False) == "transformer"
+        ):
+            mtl_mode = "transformer_transducer"
+        else:
+            mtl_mode = "transducer"
+        logging.info("Pure transducer mode")
+    elif args.mtlalpha == 1.0:
         mtl_mode = "ctc"
         logging.info("Pure CTC mode")
     elif args.mtlalpha == 0.0:
@@ -160,23 +90,26 @@ def train(args):
         mtl_mode = "mtl"
         logging.info("Multitask learning mode")
 
-    # specify model architecture
-    model_class = dynamic_import(args.model_module)
-    model = model_class(idim, odim, args)
+    if (args.enc_init is not None or args.dec_init is not None) and args.num_encs == 1:
+        model = load_trained_modules(idim_list[0], odim, args)
+    else:
+        model_class = dynamic_import(args.model_module)
+        model = model_class(
+            idim_list[0] if args.num_encs == 1 else idim_list, odim, args
+        )
     assert isinstance(model, ASRInterface)
-    subsampling_factor = model.subsample[0]
+
+    logging.info(
+        " Total parameter of the model = "
+        + str(sum(p.numel() for p in model.parameters()))
+    )
 
     if args.rnnlm is not None:
         rnnlm_args = get_model_conf(args.rnnlm, args.rnnlm_conf)
-        rnnlm = lm_pytorch.ClassifierWithState(
-            lm_pytorch.RNNLM(
-                len(args.char_list),
-                rnnlm_args.layer,
-                rnnlm_args.unit,
-                getattr(rnnlm_args, "embed_unit", None),  # for backward compatibility
-            )
+        rnnlm = default.ClassifierWithState(
+            default.RNNLM(len(args.char_list), rnnlm_args.layer, rnnlm_args.unit)
         )
-        torch.load(args.rnnlm, rnnlm)
+        torch_load(args.rnnlm, rnnlm)
         model.rnnlm = rnnlm
 
     # write model config
@@ -187,7 +120,10 @@ def train(args):
         logging.info("writing a model config file to " + model_conf)
         f.write(
             json.dumps(
-                (idim, odim, vars(args)), indent=4, ensure_ascii=False, sort_keys=True
+                (idim_list[0] if args.num_encs == 1 else idim_list, odim, vars(args)),
+                indent=4,
+                ensure_ascii=False,
+                sort_keys=True,
             ).encode("utf_8")
         )
     for key in sorted(vars(args).keys()):
@@ -203,6 +139,11 @@ def train(args):
                 % (args.batch_size, args.batch_size * args.ngpu)
             )
             args.batch_size *= args.ngpu
+        if args.num_encs > 1:
+            # TODO(ruizhili): implement data parallel for multi-encoder setup.
+            raise NotImplementedError(
+                "Data parallel is not supported for multi-encoder setup."
+            )
 
     # set torch device
     device = torch.device("cuda" if args.ngpu > 0 else "cpu")
@@ -212,21 +153,30 @@ def train(args):
         dtype = torch.float32
     model = model.to(device=device, dtype=dtype)
 
+    if args.freeze_mods:
+        model, model_params = freeze_modules(model, args.freeze_mods)
+    else:
+        model_params = model.parameters()
+
     # Setup an optimizer
     if args.opt == "adadelta":
         optimizer = torch.optim.Adadelta(
-            model.parameters(), rho=0.95, eps=args.eps, weight_decay=args.weight_decay
+            model_params, rho=0.95, eps=args.eps, weight_decay=args.weight_decay
         )
     elif args.opt == "adam":
-        optimizer = torch.optim.Adam(model.parameters(), weight_decay=args.weight_decay)
+        optimizer = torch.optim.Adam(model_params, weight_decay=args.weight_decay)
     elif args.opt == "noam":
         from espnet.nets.pytorch_backend.transformer.optimizer import get_std_opt
 
+        # For transformer-transducer, adim declaration is within the block definition.
+        # Thus, we need retrieve the most dominant value (d_hidden) for Noam scheduler.
+        if hasattr(args, "enc_block_arch") or hasattr(args, "dec_block_arch"):
+            adim = model.most_dom_dim
+        else:
+            adim = args.adim
+
         optimizer = get_std_opt(
-            model.parameters(),
-            args.adim,
-            args.transformer_warmup_steps,
-            args.transformer_lr,
+            model_params, adim, args.transformer_warmup_steps, args.transformer_lr
         )
     else:
         raise NotImplementedError("unknown optimizer: " + args.opt)
@@ -250,6 +200,12 @@ def train(args):
                 model, optimizer, opt_level=args.train_dtype
             )
         use_apex = True
+
+        from espnet.nets.pytorch_backend.ctc import CTC
+
+        amp.register_float_function(CTC, "loss_fn")
+        amp.init()
+        logging.warning("register ctc as float function")
     else:
         use_apex = False
 
@@ -258,9 +214,12 @@ def train(args):
     setattr(optimizer, "serialize", lambda s: reporter.serialize(s))
 
     # Setup a converter
-    converter = CustomConverter(
-        subsampling_factor=subsampling_factor, dtype=dtype, num_spkrs=args.num_spkrs
-    )
+    if args.num_encs == 1:
+        converter = CustomConverter(subsampling_factor=model.subsample[0], dtype=dtype)
+    else:
+        converter = CustomConverterMulEnc(
+            [i[0] for i in model.subsample_list], dtype=dtype
+        )
 
     # read json data
     with open(args.train_json, "rb") as f:
@@ -284,7 +243,7 @@ def train(args):
         batch_frames_out=args.batch_frames_out,
         batch_frames_inout=args.batch_frames_inout,
         iaxis=0,
-        oaxis=-1,
+        oaxis=0,
     )
     valid = make_batchset(
         valid_json,
@@ -299,7 +258,7 @@ def train(args):
         batch_frames_out=args.batch_frames_out,
         batch_frames_inout=args.batch_frames_inout,
         iaxis=0,
-        oaxis=-1,
+        oaxis=0,
     )
 
     load_tr = LoadInputsAndTargets(
@@ -318,30 +277,26 @@ def train(args):
     # actual bathsize is included in a list
     # default collate function converts numpy array to pytorch tensor
     # we used an empty collate function instead which returns list
-    train_iter = {
-        "main": ChainerDataLoader(
-            dataset=TransformDataset(train, lambda data: converter([load_tr(data)])),
-            batch_size=1,
-            num_workers=args.n_iter_processes,
-            shuffle=True,
-            collate_fn=lambda x: x[0],
-        )
-    }
-    valid_iter = {
-        "main": ChainerDataLoader(
-            dataset=TransformDataset(valid, lambda data: converter([load_cv(data)])),
-            batch_size=1,
-            shuffle=False,
-            collate_fn=lambda x: x[0],
-            num_workers=args.n_iter_processes,
-        )
-    }
+    train_iter = ChainerDataLoader(
+        dataset=TransformDataset(train, lambda data: converter([load_tr(data)])),
+        batch_size=1,
+        num_workers=args.n_iter_processes,
+        shuffle=not use_sortagrad,
+        collate_fn=lambda x: x[0],
+    )
+    valid_iter = ChainerDataLoader(
+        dataset=TransformDataset(valid, lambda data: converter([load_cv(data)])),
+        batch_size=1,
+        shuffle=False,
+        collate_fn=lambda x: x[0],
+        num_workers=args.n_iter_processes,
+    )
 
     # Set up a trainer
     updater = CustomUpdater(
         model,
         args.grad_clip,
-        train_iter,
+        {"main": train_iter},
         optimizer,
         device,
         args.ngpu,
@@ -363,10 +318,30 @@ def train(args):
         torch_resume(args.resume, trainer)
 
     # Evaluate the model with the test dataset for each epoch
-    trainer.extend(CustomEvaluator(model, valid_iter, reporter, device, args.ngpu))
+    if args.save_interval_iters > 0:
+        trainer.extend(
+            CustomEvaluator(model, {"main": valid_iter}, reporter, device, args.ngpu),
+            trigger=(args.save_interval_iters, "iteration"),
+        )
+    else:
+        trainer.extend(
+            CustomEvaluator(model, {"main": valid_iter}, reporter, device, args.ngpu)
+        )
 
     # Save attention weight each epoch
-    if args.num_save_attention > 0 and args.mtlalpha != 1.0:
+    is_attn_plot = (
+        (
+            "transformer" in args.model_module
+            or "conformer" in args.model_module
+            or mtl_mode in ["att", "mtl"]
+        )
+        or (
+            mtl_mode == "transducer" and getattr(args, "rnnt_mode", False) == "rnnt-att"
+        )
+        or mtl_mode == "transformer_transducer"
+    )
+
+    if args.num_save_attention > 0 and is_attn_plot:
         data = sorted(
             list(valid_json.items())[: args.num_save_attention],
             key=lambda x: int(x[1]["input"][0]["shape"][1]),
@@ -390,7 +365,42 @@ def train(args):
     else:
         att_reporter = None
 
+    # Save CTC prob at each epoch
+    if mtl_mode in ["ctc", "mtl"] and args.num_save_ctc > 0:
+        # NOTE: sort it by output lengths
+        data = sorted(
+            list(valid_json.items())[: args.num_save_ctc],
+            key=lambda x: int(x[1]["output"][0]["shape"][0]),
+            reverse=True,
+        )
+        if hasattr(model, "module"):
+            ctc_vis_fn = model.module.calculate_all_ctc_probs
+            plot_class = model.module.ctc_plot_class
+        else:
+            ctc_vis_fn = model.calculate_all_ctc_probs
+            plot_class = model.ctc_plot_class
+        ctc_reporter = plot_class(
+            ctc_vis_fn,
+            data,
+            args.outdir + "/ctc_prob",
+            converter=converter,
+            transform=load_cv,
+            device=device,
+            ikey="output",
+            iaxis=1,
+        )
+        trainer.extend(ctc_reporter, trigger=(1, "epoch"))
+    else:
+        ctc_reporter = None
+
     # Make a plot for training and validation values
+    if args.num_encs > 1:
+        report_keys_loss_ctc = [
+            "main/loss_ctc{}".format(i + 1) for i in range(model.num_encs)
+        ] + ["validation/main/loss_ctc{}".format(i + 1) for i in range(model.num_encs)]
+        report_keys_cer_ctc = [
+            "main/cer_ctc{}".format(i + 1) for i in range(model.num_encs)
+        ] + ["validation/main/cer_ctc{}".format(i + 1) for i in range(model.num_encs)]
     trainer.extend(
         extensions.PlotReport(
             [
@@ -400,7 +410,8 @@ def train(args):
                 "validation/main/loss_ctc",
                 "main/loss_att",
                 "validation/main/loss_att",
-            ],
+            ]
+            + ([] if args.num_encs == 1 else report_keys_loss_ctc),
             "epoch",
             file_name="loss.png",
         )
@@ -412,7 +423,10 @@ def train(args):
     )
     trainer.extend(
         extensions.PlotReport(
-            ["main/cer_ctc", "validation/main/cer_ctc"], "epoch", file_name="cer.png"
+            ["main/cer_ctc", "validation/main/cer_ctc"]
+            + ([] if args.num_encs == 1 else report_keys_loss_ctc),
+            "epoch",
+            file_name="cer.png",
         )
     )
 
@@ -421,14 +435,20 @@ def train(args):
         snapshot_object(model, "model.loss.best"),
         trigger=training.triggers.MinValueTrigger("validation/main/loss"),
     )
-    if mtl_mode != "ctc":
+    if mtl_mode not in ["ctc", "transducer", "transformer_transducer"]:
         trainer.extend(
             snapshot_object(model, "model.acc.best"),
             trigger=training.triggers.MaxValueTrigger("validation/main/acc"),
         )
 
     # save snapshot which contains model and optimizer states
-    trainer.extend(torch_snapshot(), trigger=(1, "epoch"))
+    if args.save_interval_iters > 0:
+        trainer.extend(
+            torch_snapshot(filename="snapshot.iter.{.updater.iteration}"),
+            trigger=(args.save_interval_iters, "iteration"),
+        )
+    else:
+        trainer.extend(torch_snapshot(), trigger=(1, "epoch"))
 
     # epsilon decay in the optimizer
     if args.opt == "adadelta":
@@ -466,6 +486,18 @@ def train(args):
                     lambda best_value, current_value: best_value < current_value,
                 ),
             )
+        # NOTE: In some cases, it may take more than one epoch for the model's loss
+        # to escape from a local minimum.
+        # Thus, restore_snapshot extension is not used here.
+        # see details in https://github.com/espnet/espnet/pull/2171
+        elif args.criterion == "loss_eps_decay_only":
+            trainer.extend(
+                adadelta_eps_decay(args.eps_decay),
+                trigger=CompareValueTrigger(
+                    "validation/main/loss",
+                    lambda best_value, current_value: best_value < current_value,
+                ),
+            )
 
     # Write a log of evaluation statistics for each epoch
     trainer.extend(
@@ -485,7 +517,7 @@ def train(args):
         "main/cer_ctc",
         "validation/main/cer_ctc",
         "elapsed_time",
-    ]
+    ] + ([] if args.num_encs == 1 else report_keys_cer_ctc + report_keys_loss_ctc)
     if args.opt == "adadelta":
         trainer.extend(
             extensions.observe_value(
@@ -511,134 +543,13 @@ def train(args):
 
     if args.tensorboard_dir is not None and args.tensorboard_dir != "":
         trainer.extend(
-            TensorboardLogger(SummaryWriter(args.tensorboard_dir), att_reporter),
+            TensorboardLogger(
+                SummaryWriter(args.tensorboard_dir),
+                att_reporter=att_reporter,
+                ctc_reporter=ctc_reporter,
+            ),
             trigger=(args.report_interval_iters, "iteration"),
         )
     # Run the training
     trainer.run()
     check_early_stop(trainer, args.epochs)
-
-
-def recog(args):
-    """Decode with the given args.
-
-    Args:
-        args (namespace): The program arguments.
-
-    """
-    set_deterministic_pytorch(args)
-    model, train_args = load_trained_model(args.model)
-    assert isinstance(model, ASRInterface)
-    model.recog_args = args
-
-    # read rnnlm
-    if args.rnnlm:
-        rnnlm_args = get_model_conf(args.rnnlm, args.rnnlm_conf)
-        if getattr(rnnlm_args, "model_module", "default") != "default":
-            raise ValueError(
-                "use '--api v2' option to decode with non-default language model"
-            )
-        rnnlm = lm_pytorch.ClassifierWithState(
-            lm_pytorch.RNNLM(
-                len(train_args.char_list),
-                rnnlm_args.layer,
-                rnnlm_args.unit,
-                getattr(rnnlm_args, "embed_unit", None),  # for backward compatibility
-            )
-        )
-        torch_load(args.rnnlm, rnnlm)
-        rnnlm.eval()
-    else:
-        rnnlm = None
-
-    if args.word_rnnlm:
-        rnnlm_args = get_model_conf(args.word_rnnlm, args.word_rnnlm_conf)
-        word_dict = rnnlm_args.char_list_dict
-        char_dict = {x: i for i, x in enumerate(train_args.char_list)}
-        word_rnnlm = lm_pytorch.ClassifierWithState(
-            lm_pytorch.RNNLM(len(word_dict), rnnlm_args.layer, rnnlm_args.unit)
-        )
-        torch_load(args.word_rnnlm, word_rnnlm)
-        word_rnnlm.eval()
-
-        if rnnlm is not None:
-            rnnlm = lm_pytorch.ClassifierWithState(
-                extlm_pytorch.MultiLevelLM(
-                    word_rnnlm.predictor, rnnlm.predictor, word_dict, char_dict
-                )
-            )
-        else:
-            rnnlm = lm_pytorch.ClassifierWithState(
-                extlm_pytorch.LookAheadWordLM(
-                    word_rnnlm.predictor, word_dict, char_dict
-                )
-            )
-
-    # gpu
-    if args.ngpu == 1:
-        gpu_id = list(range(args.ngpu))
-        logging.info("gpu id: " + str(gpu_id))
-        model.cuda()
-        if rnnlm:
-            rnnlm.cuda()
-
-    # read json data
-    with open(args.recog_json, "rb") as f:
-        js = json.load(f)["utts"]
-    new_js = {}
-
-    load_inputs_and_targets = LoadInputsAndTargets(
-        mode="asr",
-        load_output=False,
-        sort_in_input_length=False,
-        preprocess_conf=train_args.preprocess_conf
-        if args.preprocess_conf is None
-        else args.preprocess_conf,
-        preprocess_args={"train": False},
-    )
-
-    if args.batchsize == 0:
-        with torch.no_grad():
-            for idx, name in enumerate(js.keys(), 1):
-                logging.info("(%d/%d) decoding " + name, idx, len(js.keys()))
-                batch = [(name, js[name])]
-                feat = load_inputs_and_targets(batch)[0][0]
-                nbest_hyps = model.recognize(feat, args, train_args.char_list, rnnlm)
-                new_js[name] = add_results_to_json(
-                    js[name], nbest_hyps, train_args.char_list
-                )
-
-    else:
-
-        def grouper(n, iterable, fillvalue=None):
-            kargs = [iter(iterable)] * n
-            return zip_longest(*kargs, fillvalue=fillvalue)
-
-        # sort data if batchsize > 1
-        keys = list(js.keys())
-        if args.batchsize > 1:
-            feat_lens = [js[key]["input"][0]["shape"][0] for key in keys]
-            sorted_index = sorted(range(len(feat_lens)), key=lambda i: -feat_lens[i])
-            keys = [keys[i] for i in sorted_index]
-
-        with torch.no_grad():
-            for names in grouper(args.batchsize, keys, None):
-                names = [name for name in names if name]
-                batch = [(name, js[name]) for name in names]
-                feats = load_inputs_and_targets(batch)[0]
-                nbest_hyps = model.recognize_batch(
-                    feats, args, train_args.char_list, rnnlm=rnnlm
-                )
-
-                for i, name in enumerate(names):
-                    nbest_hyp = [hyp[i] for hyp in nbest_hyps]
-                    new_js[name] = add_results_to_json(
-                        js[name], nbest_hyp, train_args.char_list
-                    )
-
-    with open(args.result_label, "wb") as f:
-        f.write(
-            json.dumps(
-                {"utts": new_js}, indent=4, ensure_ascii=False, sort_keys=True
-            ).encode("utf_8")
-        )
