+++ espnet/utils/training/evaluator.py
@@ -1,12 +1,14 @@
+"""Evaluator module."""
 from chainer.training.extensions import Evaluator
 
 from espnet.utils.training.tensorboard_logger import TensorboardLogger
 
 
 class BaseEvaluator(Evaluator):
-    """Base Evaluator in ESPnet"""
+    """Base Evaluator in ESPnet."""
 
     def __call__(self, trainer=None):
+        """Call trainer and tensorboard."""
         ret = super().__call__(trainer)
         try:
             if trainer is not None:
